% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{fastCrop}
\alias{fastCrop}
\alias{fastMask}
\alias{fastRasterize}
\alias{loadPackages}
\title{Defunct functions}
\usage{
fastCrop()

fastMask()

fastRasterize()

loadPackages()
}
\description{
These functions have been removed.
}
