% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc.fixed.k.R
\name{amc.fixed.k}
\alias{amc.fixed.k}
\title{Title Adaptive Mixture Categorization}
\usage{
amc.fixed.k(data, group = 2, min_num = 1)
}
\arguments{
\item{data}{The input data for categorization}

\item{group}{The number of categories, which can be set as an integer from 2 to 10 by the user. (default = 2) We recommend the number is less than 6.}

\item{min_num}{The minimum sample size of each category, which can be set as a positive integer. (default = 1) We recommend the number is larger than 5 percent of the overall sample size.}
}
\value{
The categorized data and the categorization thresholds
\item{output}{The categorized data}
\item{cutoff}{The thresholds/cutoffs of the categorization for each variable}
}
\description{
This function is the fixed-k version of Adaptive Mixture Categorization, which categorizes variables into specified number of categories (k). The number of categories for different variables (columns) is the same.
}
