% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.amcEnv}
\alias{.amcEnv}
\title{\code{amc} package environment}
\format{An object of class \code{environment} of length 0.}
\usage{
.amcEnv
}
\description{
Environment used internally to store internal package objects.
}
\keyword{internal}
