% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-strings.R
\docType{methods}
\name{binstr}
\alias{binstr}
\title{Convert integer to binary string}
\usage{
binstr(i, maxBits = NA)
}
\arguments{
\item{i}{Positive integer <= 2^53 (<= 9.007199e+15).}

\item{maxBits}{Maximum number of bits to print (default \code{NA}).}
}
\value{
Character vector.
}
\description{
Description needed
}
\examples{
x <- sample(0:9999, 10000)
y <- binstr(x) # length is 14 bits

\dontrun{
# alternate (but slower) conversion to binary string
R.utils::intToBin(x)
}

# convert binary string to integer value (very fast)
strtoi(y, base = 2)
strtoi(substr(y, 1, 4), base = 2)
strtoi(substr(y, 5, 8), base = 2)
strtoi(substr(y, 9, 11), base = 2)
strtoi(substr(y, 12, 14), base = 2)

# see also `binary()` and `unbinary()` in the `composition` package (requires x11)

}
\author{
Alex Chubaty
}
