% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\docType{methods}
\name{cropReproj}
\alias{cropReproj}
\alias{cropReproj,RasterStack,SpatialPolygonsDataFrame-method}
\alias{cropReproj,RasterStack,Raster-method}
\alias{cropReproj,RasterLayer,ANY-method}
\alias{cropReproj,character,ANY-method}
\title{Crop and reproject a raster to a given study area}
\usage{
cropReproj(x, studyArea, ...)

\S4method{cropReproj}{RasterStack,SpatialPolygonsDataFrame}(x, studyArea,
  layerNames, filename, inRAM = FALSE, ...)

\S4method{cropReproj}{RasterStack,Raster}(x, studyArea, layerNames,
  filename, inRAM = FALSE, ...)

\S4method{cropReproj}{RasterLayer,ANY}(x, studyArea, layerNames, filename,
  inRAM = FALSE, ...)

\S4method{cropReproj}{character,ANY}(x, studyArea, layerNames, filename,
  inRAM = FALSE, ...)
}
\arguments{
\item{x}{\code{RasterStack} object or the filepath to such object.}

\item{studyArea}{A \code{SpatialPolygonsDataFrame} object.}

\item{...}{Additional arguments (not used).}

\item{layerNames}{A character vector of layer names to assign the raster.}

\item{filename}{Optional output filepath to use with \code{\link{writeRaster}}.
If not specified, a temporary file will be used.}

\item{inRAM}{Logical (default \code{FALSE}) indicating whether function
operations should be performed in memory or, if \code{TRUE},
using temporary files.}
}
\value{
\code{RasterStack} object (which can be \code{\link[raster]{unstack}}ed).
}
\description{
This function is geared toward use with study areas that are too large to work
with in memory, as it always writes temporary intermediate files to disk.
All temporary intermediate files are removed on exit to prevent "disk-full" errors,
as normally, the system temp directory is only emptied on reboot.
}
\note{
Currently, only works with a \code{RasterStack} object from file, and
is only tested using \code{studyArea} as a \code{SpatialPolygonsDataFrame}.
}
\author{
Alex Chubaty and Eliot Mcintire
}
