% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-packages.R
\docType{methods}
\name{.detach}
\alias{.detach}
\title{Detach and unload a package}
\usage{
.detach(package)
}
\arguments{
\item{package}{The name of a currently attached package.}
}
\description{
A simple wrapper to \code{detach} using \code{unload = TRUE}.
}
\seealso{
\code{\link{detach}}
}
\author{
Alex Chubaty
}
