% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{detachAllPackages}
\alias{detachAllPackages}
\title{Forcibly detach all packages}
\usage{
detachAllPackages()
}
\description{
Based on \url{https://stackoverflow.com/a/39235076/1380598}.
}
\seealso{
\code{\link{detach}}, \code{\link{detachPackage}}
}
\author{
mmfrgmpds
}
