% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{detachPackage}
\alias{detachPackage}
\title{Detach and unload a package}
\usage{
detachPackage(package)
}
\arguments{
\item{package}{The name of a currently attached package.}
}
\description{
A simple wrapper to \code{detach} using \code{unload = TRUE}.
}
\seealso{
\code{\link{detach}}, \code{\link{detachAllPackages}}
}
\author{
Alex Chubaty
}
