% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{dir.copy}
\alias{dir.copy}
\title{Copy folders with links}
\usage{
dir.copy(from, to)
}
\arguments{
\item{from}{\code{character} indicating the path to the directory to be copied.}

\item{to}{\code{character} indicating the path to which the directory will be copied.}
}
\value{
Logical indicating success or failure.
}
\description{
Copies folders like \code{file.copy} except it replicates links correctly on
unix-like systems. Based on \url{http://stackoverflow.com/a/30107868/1380598}.
}
\author{
Zach Foster

Alex Chubaty
}
