% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\docType{methods}
\name{dl.data}
\alias{dl.data}
\title{Intelligently download data}
\usage{
dl.data(urls, dest = ".", checksum = TRUE, unzip = FALSE)
}
\arguments{
\item{urls}{A character vector of data file URLs.}

\item{dest}{The directory path in which data should be downloaded.}

\item{checksum}{Logical indicating whether downloaded files should be checksummed.}

\item{unzip}{Logical indicating whether the file should be unzipped after download.}
}
\description{
Only downloads the specified files if it is not found locally.
Optionally unzips the files.
}
\author{
Alex Chubaty and Eliot Mcintire
}
