% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2raster.R, R/rasters.R
\name{dt2raster}
\alias{dt2raster}
\alias{dt2raster}
\title{Convert \code{data.table} objects to \code{RasterLayer} for plotting etc.}
\usage{
dt2raster(dt, r, val)

dt2raster(dt, r, val)
}
\arguments{
\item{dt}{A \code{data.table} object.}

\item{r}{A \code{RasterLayer} object.}

\item{val}{A string denoting data.table column name to use for raster values.}

\item{dt}{\code{data.table} object with columns \code{ID}, \code{X}, \code{Y},
and the values to assign to the raster specified by column \code{val}.}

\item{r}{\code{Raster*} object.}

\item{val}{The name of the column in \code{dt} containing the values for the raster.}
}
\value{
A \code{RasterLayer} object.
}
\description{
Convert \code{data.table} objects to \code{RasterLayer} for plotting etc.

DETAILED DESCRIPTION NEEDED
}
\author{
ALex Chubaty

Alex Chubaty
}
