% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faster-rasters.R
\name{fastCrop}
\alias{fastCrop}
\title{fastCrop}
\usage{
fastCrop(x, y, ...)
}
\arguments{
\item{x}{Raster to crop}

\item{y}{Extent object, or any object from which an Extent object can be extracted (see Details)}

\item{...}{Additional arguments as for \code{\link{writeRaster}}}
}
\description{
This function is a wrapper around velox crop function.
}
\details{
\code{fastRasterize} will try to keep the object in memory or on disk,
depending on whether the input raster was on disk.
}
