% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faster-rasters.R
\docType{methods}
\name{fastMask}
\alias{fastMask}
\alias{fastRasterize}
\title{Faster masking of a raster using a polygon}
\usage{
fastMask(stack, polygon)

fastRasterize(polygon, ras, field)
}
\arguments{
\item{stack}{A \code{RasterStack} object.}

\item{polygon}{A \code{SpatialPolygons} object.}

\item{ras}{A \code{RasterLayer} object.}

\item{field}{The field to use from \code{polygon}.}
}
\value{
A \code{Raster*} object.

A \code{Raster*} object.
}
\description{
Description needed

Description needed.
}
\note{
HAS NOT BEEN FULLY TESTED

HAS NOT BEEN FULLY TESTED
}
\examples{
\dontrun{
if (require("SpaDES", quietly = TRUE)) {
  library(raster)

  beginCluster(2)
  r <- raster(extent(0, 15, 0, 15), vals = 0)
  poly <- list()
  poly[[1]] <- SpaDES::randomPolygons(r, numTypes = 10)
  poly[[2]] <- SpaDES::randomPolygons(r, numTypes = 10)
  origStack <- stack(poly)

  shp <- origStack[[1]]
  shp[shp == shp[1]] <- NA # take a chunk out of the raster
  shp <- rasterToPolygons(shp, dissolve = TRUE) # convert to polygon

  # original mask function in raster
  newStack1 <- mask(origStack, mask = shp)

  # fastMask uses 2 clusters
  newStack2 <- fastMask(stack = origStack, polygon = shp)

  # test all equal
  identical(newStack1, newStack2)
  newStack1 <- stack(newStack1)
  newStack2 <- stack(newStack2)
  if (interactive()) {
    SpaDES::clearPlot()
    SpaDES::Plot(newStack2)
  }

  # rasterize
  shpRas1 <- rasterize(shp, origStack)
  shpRas2 <- fastRasterize(shp, origStack)
  identical(shpRas1, shpRas2)

  if (interactive()) {
    SpaDES::clearPlot()
    SpaDES::Plot(shpRas2)
  }

  endCluster()
}
}

}
\author{
Eliot Mcintire

Eliot Mcintire
}
