% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faster-rasters.R
\docType{methods}
\name{fastMask}
\alias{fastMask}
\title{Faster masking of a raster using a polygon}
\usage{
fastMask(stack, polygon)
}
\arguments{
\item{stack}{A \code{RasterStack} object.}

\item{polygon}{A \code{SpatialPolygons} object.}
}
\value{
A \code{Raster*} object.
}
\description{
Description needed
}
\note{
HAS NOT BEEN FULLY TESTED
}
\author{
Eliot Mcintire
}
