% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faster-rasters.R
\docType{methods}
\name{fastRasterize}
\alias{fastRasterize}
\title{Faster rasterizing of a polygon
memory.limit
Description needed.}
\usage{
fastRasterize(polygon, ras, field)
}
\arguments{
\item{polygon}{A \code{SpatialPolygons} object.}

\item{ras}{A \code{RasterLayer} object.}

\item{field}{The field to use from \code{polygon}.}
}
\value{
A \code{Raster*} object.
}
\description{
Faster rasterizing of a polygon
memory.limit
Description needed.
}
\note{
HAS NOT BEEN FULLY TESTED
}
\author{
Eliot Mcintire
}
