% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysmem.R
\name{.gc}
\alias{.gc}
\title{Manual garbage collection}
\usage{
.gc()
}
\description{
This shouldn't be necessary, since R (usually) handles this correctly and
automatically. However, sometimes when working with large geospatial data
(e.g., using \code{raster} and \code{sp} packages) it can help to free recently
unallocated memory manually.
}
\seealso{
\code{\link[=gc]{gc()}}
}
\author{
Alex Chubaty
}
