% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOGR.R
\docType{methods}
\name{getOGR}
\alias{getOGR}
\title{\code{readOGR} from file not in the current working directory}
\usage{
getOGR(layer, path, ...)
}
\arguments{
\item{layer}{Layer name. See \code{\link[rgdal]{readOGR}}.}

\item{path}{Parent directory of the data source.}

\item{...}{Other arguments passed to \code{readOGR}.}
}
\value{
A Spatial object.
}
\description{
Simply a wrapper for `readOGR` that allows reading a file in a directory
that is not the current working directory.
}
\author{
Alex Chubaty
}
