% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysmem.R
\docType{methods}
\name{guesstimate}
\alias{guesstimate}
\title{Guesstimate the number of CPUs for cluster operations}
\usage{
guesstimate(ram, prop = 0.8, units = "gb")
}
\arguments{
\item{ram}{How much ram is required per CPU.}

\item{prop}{Proportion of overall RAM to devote to R. Default \code{0.80}.}

\item{units}{Units of memory. One of either \code{"KB"}, \code{"MB"}, \code{"GB"}.}
}
\value{
Integer. Number of CPUs to allocate to cluster.
}
\description{
Take a wild stab at guessing how many CPUs to use in cluster when you have
some idea of how much RAM is needed per CPU.
}
\details{
Tries to be conservative by assuming no more than 80% system memory use.
}
\note{
You should take these numbers with several grains of salt.
}
\examples{
\dontrun{
guesstimate(4)
guesstimate(4, 0.90, "MB")
}

}
\author{
Alex Chubaty
}
