% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hill}
\alias{hill}
\title{Hill function}
\usage{
hill(a, b, z)
}
\arguments{
\item{a}{DESCRIPTION NEEDED}

\item{b}{DESCRIPTION NEEDED}

\item{z}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Hill function
}
\author{
Devin Goodsman
}
