% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isIntersect.R
\docType{methods}
\name{isIntersect}
\alias{isIntersect}
\title{Do spatial objects intersect?}
\usage{
isIntersect(x, layer)
}
\arguments{
\item{x}{A \code{RasterLayer} object or the name of one.}

\item{layer}{A \code{RasterLayer} object or a \code{SpatialPolygonDataFrame} object.}
}
\value{
Logical. \code{TRUE} if the intersect does occur between the layers.
}
\description{
Verifies whether the extent of two spatial objects intersect.
}
\examples{
if (require(maptools)) {
  library(raster)
  library(maptools)
  data("wrld_simpl")

  # From scratch, random raster
  xy <- matrix(rnorm(2500), 50, 50)
  inputrast <- raster(xy)

  # Set Canada extent and projection
  extent(inputrast) <- c(-170, -15, 25, 100)
  projection(inputrast) <- CRS("+proj=longlat +datum=WGS84")

  # Extract Canada boundary from wrld_simpl
  canada <- wrld_simpl[which(wrld_simpl@data$NAME == "Canada"),]

  isIntersect(inputrast, canada)
}
}
\author{
Melina Houle
}
