% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standAgeMap.R
\name{loadkNNageMap}
\alias{loadkNNageMap}
\title{Load kNN stand age map}
\usage{
loadkNNageMap(path, url = NULL, studyArea = NULL, ...)
}
\arguments{
\item{path}{file path where raster will be saved.}

\item{url}{URL from which to download the data (default provided if NULL).}

\item{studyArea}{\code{SpatialPolygonsDataFrame} giving the study area for which to extract ages.}

\item{...}{Additional arguments passed to \code{Cache} (only \code{userTags} currently used).}
}
\description{
Load kNN stand age map
}
