% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\docType{methods}
\name{geometric.mean}
\alias{geometric.mean}
\alias{harmonic.mean}
\title{Geometric and harmonic mean}
\usage{
geometric.mean(x, ...)

harmonic.mean(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Additional arguments to \code{prod} or \code{mean}.}
}
\value{
A numeric vector of length one.
}
\description{
Description needed.
}
\note{
these have not been thoroughly tested to handle \code{NA} values, etc.
}
\examples{
series <- 1:10
mean(series)
geometric.mean(series)
harmonic.mean(series)

}
\author{
Alex Chubaty
}
