% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{min_r_version}
\alias{min_r_version}
\title{Determine a package's minimum R version requirement based on its dependencies}
\usage{
min_r_version(package = NULL, exclude_main_pkg = TRUE)
}
\arguments{
\item{package}{Character string giving the name of a package whose
dependencies should be checked.}

\item{exclude_main_pkg}{Logical indicating whether \code{package} should be
excluded from the check. Default \code{TRUE}.}
}
\description{
Based on \url{https://stackoverflow.com/q/38686427}.
}
\author{
hrbrmstr and Jack Wasey
}
