% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{outerBuffer}
\alias{outerBuffer}
\title{Draw convex hull around polygons}
\usage{
outerBuffer(x, b = NULL)
}
\arguments{
\item{x}{A \code{SpatialPolygons*} object}

\item{b}{Distance to buffer.}
}
\value{
A \code{SpatialPolygons} object.
}
\description{
Draws a convex hull around vertice points of a polygon shapefile,
creating a singl polygon.
}
\author{
Ceres Barros and Alex Chubaty
}
