% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{pkgSrc}
\alias{pkgSrc}
\title{Determine source of installed packages}
\usage{
pkgSrc(pkg, lib.loc = NULL)
}
\arguments{
\item{pkg}{a character string with the package name.}

\item{lib.loc}{a character vector of directory names of \R libraries,
    or \code{NULL}.  The default value of \code{NULL} corresponds to all
    libraries currently known.  If the default is used, the loaded
    packages and namespaces are searched before the libraries.}
}
\description{
Which packages were installed from CRAN, GitHub, Bioconductor, etc.?
}
\examples{
pkgs <- as.data.frame(installed.packages(), stringsAsFactors = FALSE)
ids <- which(!(pkgs$Priority \%in\% c("base", "recommended")))
pkgs <- pkgs[ids, ]
pkgs <- pkgs$Package
pkgs[pkgSrc(pkgs) == "CRAN"]

}
