% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata-objects.R
\name{loadObjects}
\alias{loadObjects}
\alias{saveObjects}
\alias{rmObjects}
\title{Load, save, and remove .RData objects}
\usage{
loadObjects(
  objects,
  path = NULL,
  ext = ".RData",
  quiet = TRUE,
  envir = parent.frame()
)

saveObjects(
  objects,
  path = NULL,
  ext = ".RData",
  quiet = TRUE,
  envir = parent.frame()
)

rmObjects(objects, path = NULL, ext = ".RData", quiet = TRUE)
}
\arguments{
\item{objects}{A character list or character vector of object names}

\item{path}{The filepath to the directory in which to save or
from which to load the objects. The path should be
constructed using \code{\link[=file.path]{file.path()}}.}

\item{ext}{The file extension to use (default is \code{.RData}).}

\item{quiet}{Logical. Should output be suppressed? Default is \code{TRUE}.}

\item{envir}{The environment in which to look for and load objects
(default: the environment from which the function was called).}
}
\value{
Invisibly if \code{quiet=TRUE}. Either a list of objects loaded,
empty list if saved, or if removed either \code{0} for success,
\code{1} for failure.
}
\description{
Wrapper functions to \code{\link[=load]{load()}}, \code{\link[=save]{save()}},
and \code{\link[=unlink]{unlink()}}, permitting lists of objects to be
loaded/saved/deleted all at once.
}
\details{
By default, the extension \code{.RData} is used.
}
\seealso{
\code{\link[=file.path]{file.path()}}, \code{\link[=load]{load()}}, \code{\link[=save]{save()}}, \code{\link[=unlink]{unlink()}}
}
\author{
Alex Chubaty
}
