% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\docType{methods}
\name{rescale}
\alias{rescale}
\title{Rescale values to a new range}
\usage{
rescale(x, new.range = c(0, 1), old.range = NULL)
}
\arguments{
\item{x}{A numeric vector or \code{Raster*} object.}

\item{new.range}{The lower and upper bounds of the new range. Default \code{c(0,1)}.}

\item{old.range}{(optional) The lower and upper bounds of the old range.
Default \code{c(min(x), max(x)}.}
}
\value{
A new object whose values have been rescaled.
}
\description{
Adapted from \url{https://gis.stackexchange.com/a/194838/47654}.
This preserves the original distribution of the data.
}
\examples{
rescale(50, old.range = c(0, 100), new.range = c(0, 1)) ## 0.5

x <- 0:100
rescale(x) # defaults to new range [0,1]
rescale(x, c(-1, 1))

f <- system.file("external/test.grd", package = "raster")
r <- raster::raster(f)
rescale(r) # defaults to new range [0,1]
rescale(r, c(-1, 1))

}
\author{
Jeffery Evans

Alex Chubaty
}
