% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rndstr.R
\docType{methods}
\name{rndstr}
\alias{rndstr}
\title{Generate random strings}
\usage{
rndstr(n = 1, len = 8)
}
\arguments{
\item{n}{Number of strings to generate (default 1).
Will attempt to coerce to integer value.}

\item{len}{Length of strings to generate (default 8).
Will attempt to coerce to integer value.}
}
\value{
Character vector of random strings.
}
\description{
Generate a vector of random alphanumeric strings each of an arbitrary length.
}
\examples{
set.seed(11)
rndstr()
rndstr(len = 10)
rndstr(n = 5, len = 10)
rndstr(n = 5)
}
\author{
Alex Chubaty
}
