% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Simulation function}
\usage{
sim(size, nvariate, prob, mu, sigma, betay)
}
\arguments{
\item{size}{The sample size}

\item{nvariate}{The number of variable}

\item{prob}{The probability vector of distributions (probability of hidden variable)}

\item{mu}{The vector of means of normal distributions}

\item{sigma}{The vector of variances of normal distributions}

\item{betay}{Variable effects}
}
\value{
\item{state}{The hidden variable generated from a multinomial distribution}
\item{contt}{Simulated values generated from a mixture of normal distributions}
}
\description{
Simulation function
}
