% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_github.R
\name{source_github}
\alias{source_github}
\title{Source a file hosted in a pubic or private GitHub repo}
\usage{
source_github(repo, branch = "master", file, auth = Sys.getenv("GITHUB_PAT"))
}
\arguments{
\item{repo}{Name of the GitHub repository in the form \code{"user/repo"}.}

\item{branch}{Branch from which to source the file (default master).}

\item{file}{Filename to source, including relative path.}

\item{auth}{Personal Access Token to use for authorization.
Required to access files in private repositories.
By default, checks for \code{GITHUB_PAT} environment variable.
See \url{https://help.github.com/articles/creating-an-access-token-for-command-line-use/}.}
}
\description{
Source a file hosted in a pubic or private GitHub repo
}
\examples{
\dontrun{
repo = "PredictiveEcology/SpaDES"
branch = "development"
file = "_ignore/thinSpatialPolygons.R"
auth = "" ## your Personal Access Token

source_github(repo, branch, file, auth)
}
}
\author{
Alex Chubaty
}
