% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysmem.R
\docType{methods}
\name{sysmem}
\alias{sysmem}
\title{Check system memory}
\usage{
sysmem(x = "gb")
}
\arguments{
\item{x}{Units to use for output. One of either "kb", "mb", "gb".}
}
\value{
Total amount of system memory (RAM) in \code{units}.
}
\description{
This tells you the **TOTAL** system memory (RAM) available.
Other processes running on the computer will eat into this total,
and as such, you should take these numbers with a grain of salt.
}
\examples{
sysmem()

}
\author{
Alex Chubaty
}
