% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempfiles.R
\name{td}
\alias{td}
\alias{tf}
\title{Temporary directory and file creation}
\usage{
td(dir = tempdir())

tf(ext = ".tif", dir = td())
}
\arguments{
\item{dir}{Path to use as temporary directory. A subdirectory will be created.
Default is to use the R session temporary directory.}

\item{ext}{File extension to give to the newly create file.}
}
\value{
Character string indicating the filepath to the newly created file.
}
\description{
These are wrappers around \code{tempdir} and \code{tempfile} that creates the
directory or file, to ensure a correctly normalized filepath (i.e., on macOS).
}
\author{
Alex Chubaty
}
