% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempfiles.R
\docType{methods}
\name{tf}
\alias{tf}
\title{Create a temporory file}
\usage{
tf(ext = ".tif")
}
\arguments{
\item{ext}{Flie extension to give to the nwely create file.}
}
\value{
A chracter string indicating the filepath to the newly created file.
}
\description{
This is a wrapper around \code{tempfile()} that actually creates the file,
to ensure a correctly normalized filepath (i.e., on macOS)..
}
\author{
Alex Chubaty
}
