% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc-package.R
\docType{package}
\name{amc-package}
\alias{amc}
\alias{amc-package}
\title{Alex's Miscellaneous Code}
\description{
A collection of variously useful functions and utilities.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/achubaty/amc/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Eliot J B McIntire \email{eliot.mcintire@canada.ca} [contributor]
  \item Ceres Barros \email{ceresvbarros@hotmail.com} [contributor]
  \item Josh O'Brien [contributor]
  \item mmfrgmpds [contributor]
}

}
