% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc.R
\name{amc}
\alias{amc}
\title{Adaptive Mixture Categorization}
\usage{
amc(data, group = 2, min_num = 1)
}
\arguments{
\item{data}{The input data for categorization}

\item{group}{The number of categories, which can be set as an integer from 2 to 10 by the user. (default = 2)}

\item{min_num}{The minimum sample size of each category, which can be set as a positive integer. (default = 1)}
}
\value{
The categorized data and the categorization thresholds
\item{output}{The categorized data}
\item{threshold}{The thresholds/cutoffs of the categorization for each variable}
}
\description{
This is the function for the Adaptive Mixture Categorization, which can categorize variables in a flexible way using F-statistic.
}
