% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{dt2raster}
\alias{dt2raster}
\title{Convert \code{data.table} to a \code{RasterLayer} for plotting, etc.}
\usage{
dt2raster(dt, r, val)
}
\arguments{
\item{dt}{\code{data.table} object with columns \code{ID}, or both \code{X}
and \code{Y}, and the values to assign to the raster specified by
column \code{val}.}

\item{r}{\code{Raster*} object.}

\item{val}{The name of the column in \code{dt} containing the values for the raster.}
}
\value{
A \code{RasterLayer} object.
}
\description{
DETAILED DESCRIPTION NEEDED
}
\examples{
library(data.table)
library(sp)
library(raster)

r <- raster(nrows = 10, ncols = 10)
r[] <- 10

# using x,y coordinates
#dt1 <- data.table(X = , Y = , value = r[])

# using pixel ids
dt2 <- data.table(ID = 1L:ncell(r), VALUE = r[])
dt2[, VALUE := sample(1L:10L, ncell(r), replace = TRUE)]

if (interactive())
  plot(dt2raster(dt2, r, "VALUE"))

}
\author{
Alex Chubaty
}
