% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\docType{methods}
\name{dt2raster}
\alias{dt2raster}
\title{Convert \code{data.table} to a \code{RasterLayer} for plotting, etc.}
\usage{
dt2raster(dt, r, val)
}
\arguments{
\item{dt}{\code{data.table} object with columns \code{ID}, \code{X}, \code{Y},
and the values to assign to the raster specified by column \code{val}.}

\item{r}{\code{Raster*} object.}

\item{val}{The name of the column in \code{dt} containing the values for the raster.}
}
\value{
A \code{RasterLayer} object.
}
\description{
DETAILED DESCRIPTION NEEDED
}
\author{
Alex Chubaty
}
