% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysmem.R
\docType{methods}
\name{.gc}
\alias{.gc}
\title{Manual garbage collection}
\usage{
.gc()
}
\description{
This shouldn't be necessary, since R (usually) handles this correctly and
automatically. However, sometimes when working with large geospatial data
(using \code{raster} and \code{sp} packages) it can help to free recently
unallocated memory manually.
}
\seealso{
\code{\link{gc}}
}
\author{
Alex Chubaty
}
