% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deps.R
\docType{methods}
\name{get_deps}
\alias{get_deps}
\title{Get package dependencies (offline)}
\usage{
get_deps(path)
}
\arguments{
\item{path}{A local file path to a package directory.}
}
\value{
A list of package dependencies.
}
\description{
Read a package's dependencies from file, rather than searching CRAN.
From \url{http://stackoverflow.com/a/30225680/1380598}.
}
\examples{
get_deps(system.file(package = "amc"))

}
\author{
Josh O'Brien
}
